requirements.txt
# Add your requirements here like:
numpy
pandas
keras
seaborn
psycopg2
sqlalchemy
python-decouple
scikit-learn
pyflux
statsmodels
rpy2
pydot
h5py
geopandas
fbprophet
